Extracts trip answer form into reusable component
Improves maintainability and readability by refactoring
trip candidate answer UI into a dedicated component.
Removes duplicate logic from the main page and enables
easier future updates for trip answer handling.
